refactor(ui): extract page helpers into services and centralize app types
